requirements.txt
wheel==	0.33.1
PyAutoGUI==0.9.41
python-yubico==1.3.2
python3-xlib==0.15
evdev==1.1.2